+++ docs/conf.py
@@ -192,5 +192,4 @@ intersphinx_mapping = {
     'python': ('https://docs.python.org/3/', None),
     'numpy': ('https://numpy.org/doc/stable/', None),
     'xarray': ('http://xarray.pydata.org/en/stable/', None),
-    'scipy': ('https://docs.scipy.org/doc/scipy/reference/', None),
 }